refactor(design): regroup src by domain under core/ + rename hooks to handle
Two intertwined changes:

1. Domain-grouped src layout:
   packages/design/src/
     core/
       color-palette/
         apply.ts        (DOM helper: COLOR_PALETTE_STYLE_ELEMENT_ID + applyColorPaletteCss)
         bind.ts         (generic reactive wiring: bindColorPalette)
         render.ts       (pure: renderColorPaletteFromThemes)
         composable.ts   (useColorPalette + useColorPaletteUnshared)
         types.ts
         index.ts
       color-mode/
         bind.ts         (bindColorMode)
         composable.ts   (useColorMode)
         types.ts
         index.ts
       theme-runtime/
         capture.ts      (createCaptureDocument + captureColorModeAttrs)
         composable.ts   (useThemeRuntimeManager + THEME_RUNTIME_MANAGER_SYMBOL)
         types.ts
         index.ts
       index.ts
     utils/object.ts     (local isObject — keeps design Layer-0)
     index.ts

   Tests mirror at packages/design/test/unit/core/{color-palette,
   color-mode,theme-runtime}/. The old palette.spec.ts splits into
   apply.spec.ts + bind.spec.ts; the old ssr-capture.spec.ts splits
   into theme-runtime/capture.spec.ts + color-palette/render.spec.ts.

2. Theme-runtime hook method renames:
   - ColorModeHook.apply → ColorModeHook.handle
   - PaletteHook.render  → PaletteHook.handle

   Same in @vuecs/design's structural projection
   (ThemeRuntimeColorModeHook.handle / ThemeRuntimePaletteHook.handle).

   `apply` collides with `Function.prototype.apply` on every object
   method, which made the hook name ambiguous in reader context.
   `handle` is symmetric across both slots and side-steps the collision.

   Capture / render helpers also `.bind(theme.colorMode)` /
   `.bind(theme.palette)` before invoking so themes whose handle
   methods rely on `this` keep working when the function is extracted
   before call.

   theme-bootstrap, theme-bulma, theme-tailwind theme declarations
   updated. merge-themes.ts composer (defineTheme's color-mode chain)
   updated.

Public API surface unchanged (12 exports). 26 builds, 11 tests
(309 total), 0 lint errors all pass.
