--- packages/design/src/composables/index.ts
@@ -1,3 +1,0 @@
-export * from './use-color-mode';
-export * from './use-color-palette';
-export * from './use-theme-runtime';
