+++ packages/design/src/core/color-mode/index.ts
@@ -1,0 +1,3 @@
+export * from './bind';
+export * from './composable';
+export * from './types';
