+++ packages/design/src/core/color-palette/index.ts
@@ -1,0 +1,5 @@
+export * from './apply';
+export * from './bind';
+export * from './composable';
+export * from './render';
+export * from './types';
