+++ packages/design/src/core/index.ts
@@ -1,0 +1,3 @@
+export * from './color-mode';
+export * from './color-palette';
+export * from './theme-runtime';
