+++ packages/design/src/index.ts
@@ -1,7 +1,1 @@
-export * from './composables';
-export {
-    COLOR_PALETTE_STYLE_ELEMENT_ID,
-    applyColorPaletteCss,
-    bindColorPalette,
-} from './palette';
-export type { BindColorPaletteOptions, UseColorPaletteReturn } from './palette';
+export * from './core';
